+++ pybit/_websocket_stream.py
@@ -294,15 +294,12 @@ class _V5WebSocketManager(_WebSocketManager):
             "greeks",
         ]
 
-        self.symbol_wildcard = "*"
-        self.symbol_separator = "|"
-
-    def subscribe(self, topic, callback):
-        splitted_topic = topic.split(".")
-        if len(splitted_topic) > 1:
-            symbol = [splitted_topic[-1]]
-        else:
-            symbol = []
+    def subscribe(
+            self,
+            topic: str,
+            callback,
+            symbol: (str, list) = False
+    ):
 
         def prepare_subscription_args(list_of_symbols):
             """
@@ -315,9 +312,12 @@ class _V5WebSocketManager(_WebSocketManager):
                 return [topic]
 
             topics = []
-            for symbol in list_of_symbols:
-                topics.append(topic.format(symbol))
+            for single_symbol in list_of_symbols:
+                topics.append(topic.format(symbol=single_symbol))
             return topics
+
+        if type(symbol) == str:
+            symbol = [symbol]
 
         subscription_args = prepare_subscription_args(symbol)
         self._check_callback_directory(subscription_args)
@@ -333,7 +333,8 @@ class _V5WebSocketManager(_WebSocketManager):
         )
         self.ws.send(subscription_message)
         self.subscriptions[req_id] = subscription_message
-        self._set_callback(topic, callback)
+        for topic in subscription_args:
+            self._set_callback(topic, callback)
 
     def _initialise_local_data(self, topic):
         # Create self.data