+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add response headers to exceptions
 - Add argument `return_response_headers` to allow returning the response headers to the user
 - Update PyPI package's python version so that only =>3.9 is supported to prevent the error: `TypeError: 'type' object is not subscriptable`
+- Add multiple symbol support for WebSocket topics (pass `symbol` as a list)
 
 ## [5.2.0] - 2023-04-18
 ### Added